requirements.txt
h5py==2.8.0
jupyter==1.0.0
matplotlib==3.0.2
numpy==1.15.0
numpydoc==0.8.0
Pillow==5.3.0
torch==0.5.0
torchvision==0.2.1
